+++ mcfc/cudaexpression.py
@@ -17,7 +17,7 @@
 # the AUTHORS file in the main source directory for a full list of copyright
 # holders.
 
-from form import *
+from expression import *
 from cudaparameters import numElements
 
 # The ElementIndex is here and not form.py because not all backends need
